rename Shopping to AddToPackage
